pyproject.toml
[build-system]
requires = ["setuptools>=64", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "telos"
version = "0.1.0"
description = "Machine-checked congestion-control verification — one spec compiles to Lean 4 + Dafny + EBMC + Hypothesis + CPU-sim cross-checks."
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.10"
authors = [
    { name = "the telos contributors" },
]
dependencies = [
    "pydantic>=2.5",
    "PyYAML>=6.0",
    "click>=8.1",
]

[project.optional-dependencies]
hypothesis = ["hypothesis>=6.100"]
sim = ["numpy>=1.24", "matplotlib>=3.7"]
dev = [
    "pytest>=7.4",
    "hypothesis>=6.100",
    "numpy>=1.24",
    "matplotlib>=3.7",
]

[project.scripts]
telos = "telos.cli:main"

[tool.setuptools.packages.find]
include = ["telos*"]
exclude = ["tests*", "examples*", "docs*"]
